+++ src/main/java/com/woowacourse/MatzipApplication.java
@@ -1,4 +1,4 @@
-package com.woowacourse.matzip;
+package com.woowacourse;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
