+++ src/test/java/com/woowacourse/MatzipApplicationTests.java
@@ -1,7 +1,6 @@
-package com.woowacourse.matzip;
+package com.woowacourse;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
 class MatzipApplicationTests {
